Revert "[RUBRICA] [ANTES DA T7]"
