+++ .eslintignore
@@ -1,4 +1,3 @@
 /tests
 /migrations/
-/seeders/
 /models/index.js
