--- controllers/index.js
@@ -1,11 +1,0 @@
-const loginController = require('./login.controller');
-const postController = require('./post.controller');
-const userController = require('./user.controller');
-const categoriesController = require('./categories.controller');
-
-module.exports = {
-  postController,
-  loginController,
-  userController,
-  categoriesController,
-};
