--- models/Categories.js
@@ -1,11 +1,0 @@
-module.exports = (sequelize, DataTypes) => {
-  const Category = sequelize.define(
-    'Categories',
-    {
-      name: DataTypes.STRING,
-    },
-    { timestamps: false },
-  );
-
-  return Category;
-};
