+++ package.json
@@ -3,17 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@hapi/joi": "^17.1.1",
-    "@types/jest": "25.1.4",
-    "body-parser": "^1.19.0",
-    "dotenv": "^8.2.0",
-    "eslint-config-trybe-backend": "1.0.3",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "mysql2": "^2.1.0",
-    "sequelize": "^6.3.4",
-    "shelljs": "^0.8.4",
-    "wait-on": "^5.3.0"
+    "dotenv": "8.2.0",
+    "eslint-config-trybe-backend": "^1.0.3",
+    "express": "4.17.1",
+    "mysql2": "2.1.0",
+    "sequelize": "6.3.4",
+    "shelljs": "0.8.4"
   },
   "scripts": {
     "prestart": "npx sequelize-cli db:create && npx sequelize-cli db:migrate $",
@@ -23,12 +18,13 @@
     "test": "jest --runInBand",
     "test-coverage": "jest --runInBand --coverage",
     "debug": "nodemon index.js",
-    "lint": "eslint --no-inline-config --ext .js,.jsx --no-error-on-unmatched-pattern -c .eslintrc.json ."
+    "lint": "eslint --no-inline-config --no-error-on-unmatched-pattern -c .eslintrc.json ."
   },
   "devDependencies": {
     "eslint": "6.8.0",
     "eslint-config-airbnb-base": "14.2.0",
-    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-import": "2.21.2",
+    "eslint-plugin-sonarjs": "0.5.0",
     "frisby": "2.1.2",
     "jest": "26.0.1",
     "sequelize-cli": "6.2.0",
