--- seeders/20200812184236-Categories.js
@@ -1,19 +1,0 @@
-module.exports = {
-  up: async (queryInterface, _Sequelize) => {
-    await queryInterface.bulkInsert('Categories',
-      [
-        {
-          id: 1,
-          name: 'Inovação',
-        },
-        {
-          id: 2,
-          name: 'Escola',
-        },
-      ], { timestamps: false });
-  },
-
-  down: async (queryInterface, _Sequelize) => {
-    await queryInterface.bulkDelete('Categories', null, {});
-  },
-};
