--- seeders/20210430010915-postCategories.js
@@ -1,20 +1,0 @@
-module.exports = {
-  up: async (queryInterface, _Sequelize) => {
-    await queryInterface.bulkInsert('postsCategories',
-      [
-        {
-          postId: 1,
-          categoryId: 1,
-        },
-        {
-          postId: 2,
-          categoryId: 2,
-        },
-
-      ], { timestamps: false });
-  },
-
-  down: async (queryInterface, _Sequelize) => {
-    await queryInterface.bulkDelete('Categories', null, {});
-  },
-};
