--- services/index.js
@@ -1,11 +1,0 @@
-const loginServices = require('./login.services');
-const postServices = require('./post.services');
-const userServices = require('./user.services');
-const categoriesServices = require('./categories.services');
-
-module.exports = {
-  loginServices,
-  postServices,
-  userServices,
-  categoriesServices,
-};
