+++ tests/login.test.js
@@ -94,7 +94,7 @@ describe('2 - Sua aplicação deve ter o endpoint POST `/login`', () => {
       .then((response) => {
         const { body } = response;
         const result = JSON.parse(body);
-        expect(result.message).toBe('Invalid fields');
+        expect(result.message).toBe('Campos inválidos');
       });
   });
 });
