+++ tests/searchPost.test.js
@@ -3,7 +3,7 @@ const shell = require('shelljs');
 
 const url = 'http://localhost:3000';
 
-describe('13 - Sua aplicação deve ter o endpoint GET `post/search?q=:searchTerm`', () => {
+describe('10 - Sua aplicação deve ter o endpoint GET `post/search?q=:searchTerm`', () => {
   beforeEach(() => {
     shell.exec('npx sequelize-cli db:drop $');
     shell.exec('npx sequelize-cli db:create && npx sequelize-cli db:migrate $');
@@ -48,8 +48,6 @@ describe('13 - Sua aplicação deve ter o endpoint GET `post/search?q=:searchTer
         expect(result[0].user.displayName).toBe('Lewis Hamilton');
         expect(result[0].user.email).toBe('lewishamilton@gmail.com');
         expect(result[0].user.image).toBe('https://upload.wikimedia.org/wikipedia/commons/1/18/Lewis_Hamilton_2016_Malaysia_2.jpg');
-        expect(result[0].categories[0].id).toBe(2);
-        expect(result[0].categories[0].name).toBe("Escola");
       });
   });
 
@@ -195,7 +193,7 @@ describe('13 - Sua aplicação deve ter o endpoint GET `post/search?q=:searchTer
       .then((response) => {
         const { body } = response;
         const result = JSON.parse(body);
-        expect(result.message).toBe('Token not found');
+        expect(result.message).toBe('Token não encontrado');
       });
   });
 
@@ -214,7 +212,7 @@ describe('13 - Sua aplicação deve ter o endpoint GET `post/search?q=:searchTer
       .then((response) => {
         const { body } = response;
         const result = JSON.parse(body);
-        expect(result.message).toBe('Expired or invalid token');
+        expect(result.message).toBe('Token expirado ou inválido');
       });
   });
 });
